requirements.txt
certifi==2024.8.30
packaging==24.1
pygame==2.6.1
openai==1.12.0
python-dotenv==1.0.0
PyOpenGL==3.1.6
numpy>=1.24.1
websockets==12.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.1
pydub==0.25.1
python-socketio==5.11.1
